move error checks
