+++ stac_fastapi/mongo/stac_fastapi/mongo/types/error_checks.py
